[skip CI] More refactoring
